property-recommendation/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.3.1
python-dotenv==1.0.0
scikit-learn==1.4.0
pandas==2.1.3
numpy==1.25.2
python-multipart==0.0.6